+++ docs/changelog.rst
@@ -17,6 +17,8 @@ Additions to the :doc:`core-object-methods`.
 
 * Added the :py:meth:`khoros.core.Khoros.Board.get_message_count` method.
 * Added the :py:meth:`khoros.core.Khoros.Board.get_all_messages` method.
+* Added the :py:class:`khoros.core.Khoros.Label` class with the following methods:
+    * :py:meth:`khoros.core.Khoros.Label.get_labels_for_message`
 * Added the ability to define the logging level when instantiating the core object.
 * Added additional log messages to the core object.
 
@@ -28,6 +30,7 @@ Additions to the :doc:`primary modules <primary-modules>`.
 * Added the :py:func:`khoros.structures.boards.get_all_messages` function.
 * Added the :py:func:`khoros.structures.boards._perform_single_query` function.
 * Added the :py:func:`khoros.structures.boards._add_missing_cols` function.
+* Added the :py:func:`khoros.objects.labels.get_labels_for_message` function.
 
 Changed
 =======